% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumB.r
\name{cumB}
\alias{cumB}
\title{cumB}
\usage{
cumB(x, y, npoints = 5000)
}
\arguments{
\item{x}{Descrizione del parametro x.}
}
\value{
Descrizione di ciò che viene restituito.
}
\description{
Descrizione dettagliata della funzione.
}
