% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumBdata.r
\name{cumBdata}
\alias{cumBdata}
\title{Calculate Cumulative Biomass}
\usage{
cumBdata(data, TL_col, B_col, threshold = 2.4)
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{TL_col}{A character string specifying the column name representing TL.}

\item{B_col}{A character string specifying the column name representing B.}

\item{threshold}{A numeric value to filter rows based on the TL column.}
}
\value{
A dataframe with rows filtered, arranged, and with columns `cumB` and `cumBst` added.
`cumB` contains the cumulative values of the B column, and `cumBst` contains the standardized cumulative values of the B column.
#' @importFrom dplyr filter arrange mutate
#' @importFrom ggplot2 filter arrange mutate
}
\description{
This function calculates the cumulative biomass for given data, filtering rows based on a threshold
for the TL column, and then arranging and creating cumulative values.
}
\examples{
\dontrun{
dat <- data.frame(TL = c(1, 2.5, 3, 4), B = c(10, 20, 30, 40))
result <- cumulative_B(dat, "TL", "B", 2.4)
}
}
