% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumBfit_raw.r
\name{cumBfit_raw}
\alias{cumBfit_raw}
\title{cumBfit}
\usage{
cumBfit_raw(cumBdata, npoints = 5000)
}
\arguments{
\item{cumBdata}{Dataset}

\item{y}{Descrizione del parametro y.}
}
\value{
Descrizione di ciò che viene restituito.
}
\description{
Descrizione dettagliata della funzione.
}
