% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venicelagoon.r
\docType{data}
\name{venicelagoon}
\alias{venicelagoon}
\title{Titolo del Dataset: Venice lagoon food web}
\format{
Un dataframe con x righe e y colonne:
\describe{
  \item{var1}{Descrizione della variabile var1.}
  \item{var2}{Descrizione della variabile var2.}
  ... (e così via per ogni variabile nel dataframe) ...
}
}
\source{
Anelli Monti et al., 2019
}
\usage{
venicelagoon
}
\description{
Breve descrizione del dataset.
}
\examples{
\dontrun{
data(venicelagoon)
head(venicelagoon)
}
}
\references{
Qualsiasi riferimento collegato al dataset.
}
\keyword{datasets}
