% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumBfit.r
\name{cumBfit}
\alias{cumBfit}
\title{cumBfit}
\usage{
cumBfit(
  x,
  y,
  npoints = 2000,
  drmfun = baro5(fixed = c(NA, NA, NA, 1, NA)),
  TLrange = c(1.6, 5)
)
}
\arguments{
\item{x}{Descrizione del parametro x.}

\item{y}{Descrizione del parametro y.}
}
\value{
Descrizione di ciò che viene restituito.
}
\description{
Descrizione dettagliata della funzione.
}
