% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumBfit.r
\name{cumBfit}
\alias{cumBfit}
\title{cumBfit}
\usage{
cumBfit(x, y, npoints = 5000)
}
\arguments{
\item{x}{Descrizione del parametro x.}

\item{y}{Descrizione del parametro y.}
}
\value{
Descrizione di ciò che viene restituito.
}
\description{
Descrizione dettagliata della funzione.
}
